fix(rankings-loader): handle duplicate teams and ranking entries
